Created Enemy Fish Obstacle
An obstacle that functions the same way as Box Obstacle. Probably should
be defined as a Static Body type. Call setGoal when you create the fish.
Call moveTowardsGoal() during update to get the fish to move from its
current position to the goal state. Call Patrol() in update if you want
a fish to move between two different points.